Clean-up: - Refactored NonRepairable into non_repairable.py - Organised imports - Made flake8 happy - Removed unnecessary __init__.py file - Moved helper classes into their own file
