--- repyability/rbd/__init__.py
@@ -1,1 +1,0 @@
-from .rbd import RBD
